requirements.txt
pytorch-lightning==2.0.8
resemblyzer
g2p-en
inflect
librosa
matplotlib
numpy
numba
pypinyin
pyworld
PyYAML
scikit-learn
scipy
soundfile
tensorboard
tgt
torch
tqdm
unidecode
torchaudio
pandas
xmltodict
jiwer
praatio
comet-ml
datasets
flashlight-text
